Added View Balance Feature and login was handled through servlet
