+++ src/main/java/in/mohamedhalith/service/EmployeeManager.java
@@ -34,4 +34,22 @@ public class EmployeeManager {
 		return employeeList;
 	}
 
+	/**
+	 * This method is used to return a specific employee. Employee username is
+	 * obtained and returns employee with that username.
+	 * 
+	 * @param username
+	 * @return
+	 */
+	public static Employee getEmployee(String username) {
+		Employee getEmployee = null;
+		for (Employee employee : employeeList) {
+			if (employee.getUsername().equals(username)) {
+				getEmployee = employee;
+				break;
+			}
+		}
+		return getEmployee;
+	}
+
 }
