+++ src/main/java/in/mohamedhalith/servlet/LoginServlet.java
@@ -7,6 +7,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import in.mohamedhalith.validator.LoginValidator;
 
@@ -32,11 +33,15 @@ public class LoginServlet extends HttpServlet {
 			boolean admin = LoginValidator.adminVerification(username, password, role);
 			boolean employee = LoginValidator.employeeVerification(username, password, role);
 			if (admin) {
-				message = "Admin Logged In Successfully";
-				response.sendRedirect("login.jsp?infoMessage=" + message);
+				HttpSession session = request.getSession();
+				session.setAttribute("ROLE", role);
+				session.setAttribute("LOGGEDIN_USERNAME", username);
+				response.sendRedirect("index.jsp");
 			} else if (employee) {
-				message = "Employee Logged In Successfully";
-				response.sendRedirect("login.jsp?infoMessage=" + message);
+				HttpSession session = request.getSession();
+				session.setAttribute("ROLE", role);
+				session.setAttribute("LOGGEDIN_USERNAME", username);
+				response.sendRedirect("index.jsp");
 			} else {
 				message = "Invalid Credentials";
 				response.sendRedirect("login.jsp?infoMessage=" + message);
