+++ src/test/java/in/mohamedhalith/service/TestDisplayEmployees.java
@@ -14,7 +14,28 @@ public class TestDisplayEmployees {
 	@Test
 	public void testGetEmployeeList() {
 		List<Employee> employeeList = EmployeeManager.getEmployeeList();
-		assertEquals(2,employeeList.size());
+		assertEquals(2, employeeList.size());
 	}
 
+	@Test
+	public void testGetEmployeeWithFirstEmployee() {
+		String username = "moha2627";
+		Employee employee = EmployeeManager.getEmployee(username);
+		assertEquals("Mohamed", employee.getName());
+		assertEquals("2627moha", employee.getPassword());
+		assertEquals(2, employee.getSickLeave());
+		assertEquals(2, employee.getCasualLeave());
+		assertEquals(1, employee.getEarnedLeave());
+	}
+
+	@Test
+	public void testGetEmployeeWithSecondEployee() {
+		String username = "hali2628";
+		Employee employee = EmployeeManager.getEmployee(username);
+		assertEquals("Halith", employee.getName());
+		assertEquals("2628hali", employee.getPassword());
+		assertEquals(2, employee.getSickLeave());
+		assertEquals(2, employee.getCasualLeave());
+		assertEquals(1, employee.getEarnedLeave());
+	}
 }
